pyproject.toml
[project]
name = "horizon"
version = "0.1.0"
description = "AI-driven information aggregation system for tracking academic and social trends"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "feedparser>=6.0.11",
    "beautifulsoup4>=4.12.3",
    "anthropic>=0.39.0",
    "openai>=1.54.0",
    "google-genai>=0.3.0",
    "pydantic>=2.9.0",
    "rich>=13.9.0",
    "tenacity>=9.0.0",
    "python-dotenv>=1.0.0",
    "ddgs>=7.0.0",
    "markdown>=3.10.2",
    "mcp>=1.0.0",
    "supabase>=2.3.0",
    "fastapi>=0.136.1",
    "playwright>=1.54.0",
    "uvicorn>=0.41.0",
    "telethon>=1.33.0",
]

[project.scripts]
horizon = "src.cli:main"
horizon-mcp = "src.mcp.server:main"
horizon-wizard = "src.configuration.setup_wizard:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
